machine: Work on the old implementation
The second implementation of the machine wasn't doing it for me, so I'm
going back to a modified version of the first implementation.

I noticed that the old implementation of Where was wrong (it didn't
evaluate the expressions in the binding part), so I removed it and
wrote a new test case. I'll write a new implementation soon.